store sum and number of votes in each quote
This seems to improve the response time.  We don't store a list of
votes for each quote, which mean we don't have to retrieve every vote
when we get the list of quotes.  The only time we need to access the
vote table is when we add a new vote.